+++ src/Entity/Utils/ResultOfCalcStorageFee.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Extraton\TonClient\Entity\Utils;
+
+use Extraton\TonClient\Entity\AbstractResult;
+
+/**
+ * Type ResultOfCalcStorageFee
+ */
+class ResultOfCalcStorageFee extends AbstractResult
+{
+    /**
+     * Get fee
+     *
+     * @return string
+     */
+    public function getFee(): string
+    {
+        return $this->requireString('fee');
+    }
+}